testing user file copy
